fixed size_t errors for windows users, updated publication

+++ edstem/integration/storage/bh.cpp
@@ -82,7 +82,7 @@ void BinaryHeap::printBinaryHeap(int filter, int specifier, std::vector<std::str
             else if (filter == 2 && heap[i]->data.second.first[1] != "Unavailable") {
                 checkCol(i, j);
             }
-            else if (filter == 3 && heap[i]->data.second.first.size() > 6) {
+            else if (filter == 3 && (int)heap[i]->data.second.first.size() > 6) {
                 checkCol(i, j);
             }
             else if (filter == 4 && specifier == 1 && heap[i]->data.second.first[3] == "True") {
@@ -240,12 +240,12 @@ void BinaryHeap::printStatus(int i, int j){
 }
 
 void BinaryHeap::printUserScore(int i, int j){
-    std::cout << RED << "User Score: " << RESET << heap[i]->data.second.first[heap[i]->data.second.first.size() -3]  << " ";
+    std::cout << RED << "User Score: " << RESET << heap[i]->data.second.first[(int)heap[i]->data.second.first.size() -3]  << " ";
     return;
 }
 
 void BinaryHeap::printPotentialScore(int i, int j){
-    std::cout << RED << "Potential Score: " << RESET << heap[i]->data.second.first[heap[i]->data.second.first.size() -2]  << " ";
+    std::cout << RED << "Potential Score: " << RESET << heap[i]->data.second.first[(int)heap[i]->data.second.first.size() -2]  << " ";
     return;
 }
 
@@ -290,7 +290,7 @@ void BinaryHeap::printKeyFront(int i, int filter, int specifier, std::vector<std
     else if (filter == 2 && heap[i]->data.second.first[1] != "Unavailable") {
         std::cout << "Key: " << GREEN << heap[i]->data.first << RESET << " [";
     } 
-    else if (filter == 3 && heap[i]->data.second.first.size() > 6) {
+    else if (filter == 3 && (int)heap[i]->data.second.first.size() > 6) {
         std::cout << "Key: " << GREEN << heap[i]->data.first << RESET << " [";
     } 
     else if (filter == 4 && specifier == 1 && heap[i]->data.second.first[3] == "True") {
@@ -321,7 +321,7 @@ void BinaryHeap::printKeyBack(int i, int filter, int specifier, std::vector<std:
     else if (filter == 2 && heap[i]->data.second.first[1] != "Unavailable") {
         std::cout << "]" << std::endl << std::endl;
     } 
-    else if (filter == 3 && heap[i]->data.second.first.size() > 6) {
+    else if (filter == 3 && (int)heap[i]->data.second.first.size() > 6) {
         std::cout << "]" << std::endl << std::endl;
     } 
     else if (filter == 4 && specifier == 1 && heap[i]->data.second.first[3] == "True") {
